# swagger/api/apiauth.yaml
openapi: "3.0.3"
info:
  title: AuthenticateMe - an API for authentication
  description: "This is an Authorization server.  
  For sample, you can use the api key \"special-key\" to test the authentication"
  version: 1.0.0
  termsOfService: "http://auth.io/terms/"
  contact:
    email: "authMe@authm.io"
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: https://{environment}.authenticate.com/
    variables:
      environment:
        default: api    # Production server
        enum:
          - api         # Production server
          - api.dev     # Development server
          - api.staging # Staging server
security:
  - basicAuth: []
    keyAuth: []
paths:
...
  /auth/deactivate:
    post:
      summary: Deactivate user
      description:  Deactivation for user not to be able to login, can be triggered either by user himself, either by "administrator" user or process.
      operationId: deactivateUser
      tags:
        - users

      parameters: 
        - in: header
          name: Authorization
          description: JWT is used to know if the user is authorized and allowed to perform the request
          schema:
            type: string
          required: true

      requestBody:
        description: information required for searching the user , for regular user only his login will be accepted, for admin user - any login from DB 
        required: true
        content: 
          application/json:
            schema: 
              $ref: '#/components/schemas/DeactivationRequest' 
            example:
              login: johnny
      
      responses:
        "200":
          description: deactivated cuccessfully 
        "400":
          description: Invalid request body  
        "401":
          $ref: '#/components/responses/UnauthorizedErr' 
        "403":
          description: User has no rights to perform operation
        "404":
          description: Invalid login - doesn`t exist in DB
        "406":
          description: User is deactivated already
        "500":
          description: Internal server error
...
components:
  securitySchemes:
    basicAuth:
      type: http
      description: Basic http authentication
      scheme: basic
    keyAuth:
      type: apiKey
      description: API key to authorize requests.
      in: header
      name: api_key
  responses:
    UnauthorizedErr:
      description: API key and/or basicAuth are invalid or missing
      headers:
        WWW_Authenticate:
          description: Unauthorized
          schema:
            type: string
...
  schemas:
...
    DeactivationRequest:
      type: object
      properties:
        login:
          type: string
      required:
          - login
      example:
          login : johnny     
...